Move enum conversion to their own files
Each enum defined in electrical_components/_xxx.py now has its own
converter module at proto/v1alpha8/_xxx.py and test file:

- _category.py: electrical_component_category_{from,to}_proto
- _state_code.py: electrical_component_state_code_{from,to}_proto
- _diagnostic_code.py: electrical_component_diagnostic_code_{from,to}_proto
- test_category.py: TestElectricalComponentCategoryParity
- test_state_code.py: TestElectricalComponentStateCodeParity
- test_diagnostic_code.py: TestElectricalComponentDiagnosticCodeParity

This mirrors the domain layout convention where each public type lives
in its own underscore-prefixed module, rather than bundling unrelated
enum converters together in _electrical_component.py.
