+++ src/frequenz/client/common/microgrid/electrical_components/proto/v1alpha8/__init__.py
@@ -3,19 +3,25 @@
 
 """Conversion of electrical component enums from/to protobuf v1alpha8."""
 
-from ._electrical_component import (
+from ._category import (
     electrical_component_category_from_proto,
     electrical_component_category_to_proto,
+)
+from ._diagnostic_code import (
     electrical_component_diagnostic_code_from_proto,
     electrical_component_diagnostic_code_to_proto,
+)
+from ._electrical_component import (
     electrical_component_from_proto,
     electrical_component_from_proto_with_issues,
-    electrical_component_state_code_from_proto,
-    electrical_component_state_code_to_proto,
 )
 from ._electrical_component_connection import (
     electrical_component_connection_from_proto,
     electrical_component_connection_from_proto_with_issues,
+)
+from ._state_code import (
+    electrical_component_state_code_from_proto,
+    electrical_component_state_code_to_proto,
 )
 
 __all__ = [
