+++ tests/microgrid/electrical_components/proto/v1alpha8/test_state_code.py
@@ -1,0 +1,27 @@
+# License: MIT
+# Copyright © 2026 Frequenz Energy-as-a-Service GmbH
+
+"""Tests for electrical component state code to/from protobuf v1alpha8 conversion."""
+
+from frequenz.api.common.v1alpha8.microgrid.electrical_components import (
+    electrical_components_pb2,
+)
+
+from frequenz.client.common.microgrid.electrical_components import (
+    ElectricalComponentStateCode,
+)
+from frequenz.client.common.microgrid.electrical_components.proto.v1alpha8 import (
+    electrical_component_state_code_from_proto,
+    electrical_component_state_code_to_proto,
+)
+from frequenz.client.common.test.enum_parity import EnumParityTest
+
+
+class TestElectricalComponentStateCodeParity(EnumParityTest):
+    """Parity tests for the `ElectricalComponentStateCode` enum."""
+
+    python_enum = ElectricalComponentStateCode
+    proto_enum = electrical_components_pb2.ElectricalComponentStateCode
+    name_prefix = "ELECTRICAL_COMPONENT_STATE_CODE_"
+    from_proto = staticmethod(electrical_component_state_code_from_proto)
+    to_proto = staticmethod(electrical_component_state_code_to_proto)
